Convert window size values to usize
The window size was being provided as it's native type of an `i32`,
where it was often used with `usize` values. This change converts the
`i32` to a `usize` on read so that the conversion was not needed when it
was used. Since the window size should easily fit inside a `usize` on
any system that would be running git, this should be fine.

+++ src/constants.rs
@@ -1,8 +1,8 @@
 pub(crate) const TITLE: &str = "Git Interactive Rebase Tool";
-pub(crate) const TITLE_LENGTH: i32 = 27;
+pub(crate) const TITLE_LENGTH: usize = 27;
 pub(crate) const TITLE_SHORT: &str = "Git Rebase";
-pub(crate) const TITLE_SHORT_LENGTH: i32 = 10;
-pub(crate) const TITLE_HELP_INDICATOR_LENGTH: i32 = 6;
+pub(crate) const TITLE_SHORT_LENGTH: usize = 10;
+pub(crate) const TITLE_HELP_INDICATOR_LENGTH: usize = 6;
 
 pub(crate) const HEIGHT_ERROR_MESSAGE: &str = "Window too small, increase height to continue\n";
 pub(crate) const MINIMUM_WINDOW_HEIGHT_ERROR_WIDTH: usize = 45;
